mainloging/microservices/mainloging/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
pwdlib[argon2,bcrypt]
pydantic
python-dotenv
psycopg2-binary
email-validator
anyio
starlette
watchfiles